LIME-450: * Refactored constants on profile setting to add type error on Notification manager
