+++ frontend/src/bundles/profile/components/profile-settings/constants/default-update-profile-payload.ts
@@ -1,0 +1,15 @@
+import { Gender } from '~/bundles/common/enums/enums.js';
+import { type UserUpdateProfileRequestDto } from '~/bundles/users/users.js';
+
+const DEFAULT_UPDATE_PROFILE_PAYLOAD: UserUpdateProfileRequestDto = {
+    avatarUrl: '',
+    fullName: '',
+    username: '',
+    dateOfBirth: '',
+    weight: '',
+    height: '',
+    location: '',
+    gender: Gender.MALE,
+};
+
+export { DEFAULT_UPDATE_PROFILE_PAYLOAD };
