+++ frontend/src/bundles/profile/components/profile-settings/constants/error-wrong-filetype-img.ts
@@ -1,0 +1,4 @@
+const ERROR_WRONG_FILETYPE_IMG =
+    'The file you uploaded is not an image. Please upload a valid image file.';
+
+export { ERROR_WRONG_FILETYPE_IMG };
